requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
slowapi
psycopg[binary]==3.*
psycopg-pool==3.*